diako:تغییرات پیشخوان انجام شد

+++ app/src/main/java/ir/tdaapp/paymanyar/Model/Services/S_LevelFragment.java
@@ -4,7 +4,7 @@ import android.graphics.Bitmap;
 
 public interface S_LevelFragment {
     void OnStart();
-    void onPositionBubble(int x, int y);
+    void onPositionBubble(float x, float y);
     Float onPixelTodp(float pixel);
     Float onDpTopixel(float dp);
     void onCenterBubble(boolean isCenter);
